+++ base/database/batch_test.go
@@ -16,7 +16,7 @@ var defaultValues = TestTableSlice{
 	{Name: "M", Email: "N"},
 }
 
-//nolint: errcheck
+// nolint: errcheck
 func TestBatchInsert(t *testing.T) {
 	utils.SkipWithoutDB(t)
 	Configure()
