+++ base/types/inventory/inventory.go
