+++ evaluator/evaluate.go
@@ -8,8 +8,8 @@ import (
 	"app/base/models"
 	"app/base/mqueue"
 	"app/base/types"
+	"app/base/types/vmaas"
 	"app/base/utils"
-	"app/base/vmaas"
 	"context"
 	"encoding/json"
 	"net/http"
