+++ evaluator/evaluate_packages.go
@@ -3,8 +3,8 @@ package evaluator
 import (
 	"app/base/database"
 	"app/base/models"
+	"app/base/types/vmaas"
 	"app/base/utils"
-	"app/base/vmaas"
 	"bytes"
 	"encoding/json"
 	"strings"
