+++ evaluator/evaluate_test.go
@@ -5,8 +5,8 @@ import (
 	"app/base/database"
 	"app/base/models"
 	"app/base/mqueue"
+	"app/base/types/vmaas"
 	"app/base/utils"
-	"app/base/vmaas"
 	"context"
 	"net/http"
 	"os"
