+++ listener/upload.go
@@ -4,12 +4,12 @@ import (
 	"app/base"
 	"app/base/api"
 	"app/base/database"
-	"app/base/inventory"
 	"app/base/models"
 	"app/base/mqueue"
 	"app/base/types"
+	"app/base/types/inventory"
+	"app/base/types/vmaas"
 	"app/base/utils"
-	"app/base/vmaas"
 	"app/manager/middlewares"
 	"crypto/sha256"
 	"encoding/hex"
