+++ listener/upload_test.go
@@ -4,12 +4,12 @@ import (
 	"app/base/api"
 	"app/base/core"
 	"app/base/database"
-	"app/base/inventory"
 	"app/base/models"
 	"app/base/mqueue"
 	"app/base/types"
+	"app/base/types/inventory"
+	"app/base/types/vmaas"
 	"app/base/utils"
-	"app/base/vmaas"
 	"context"
 	"errors"
 	"fmt"
