+++ manager/controllers/package_versions_test.go
@@ -8,7 +8,7 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-//nolint: dupl
+// nolint: dupl
 func TestPackageVersions(t *testing.T) {
 	core.SetupTest(t)
 	w := CreateRequestRouterWithParams("GET", "/firefox/versions", nil, "", PackageVersionsListHandler, 3,
