+++ tasks/vmaas_sync/advisory_sync_test.go
@@ -5,8 +5,8 @@ import (
 	"app/base/database"
 	"app/base/models"
 	"app/base/types"
+	"app/base/types/vmaas"
 	"app/base/utils"
-	"app/base/vmaas"
 	"encoding/json"
 	"testing"
 	"time"
