+++ tasks/vmaas_sync/package_sync.go
@@ -4,8 +4,8 @@ import (
 	"app/base"
 	"app/base/database"
 	"app/base/models"
+	"app/base/types/vmaas"
 	"app/base/utils"
-	"app/base/vmaas"
 	"crypto/sha256"
 	"net/http"
 	"time"
