+++ tasks/vmaas_sync/repo_based.go
@@ -4,8 +4,8 @@ import (
 	"app/base"
 	"app/base/database"
 	"app/base/mqueue"
+	"app/base/types/vmaas"
 	"app/base/utils"
-	"app/base/vmaas"
 	"net/http"
 	"time"
 )
