+++ tasks/vmaas_sync/vmaas_sync_test.go
@@ -6,8 +6,8 @@ import (
 	"app/base/database"
 	"app/base/models"
 	"app/base/mqueue"
+	"app/base/types/vmaas"
 	"app/base/utils"
-	"app/base/vmaas"
 	"app/tasks"
 	"context"
 	"net/http"
